Fully decouple networking component
