+++ Sources/LiveKit/Broadcast/BroadcastMessage.swift
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2025 LiveKit
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#if os(iOS)
+
+import Foundation
+
+/// A message sent between the broadcast extension and main app.
+enum BroadcastMessage: Codable {
+    case imageSample(BroadcastImageSample)
+}
+
+extension BroadcastMessage {
+    
+    func transportEncoded() throws -> Data {
+        return try Self.encoder.encode(self)
+    }
+    
+    init(transportEncoded data: Data) throws {
+        self = try Self.decoder.decode(Self.self, from: data)
+    }
+    
+    private static let encoder = PropertyListEncoder()
+    private static let decoder = PropertyListDecoder()
+}
+
+#endif
